i did what i can and i wanna commit game end
