--- PasswordDatabase.java
@@ -1,38 +1,0 @@
-package PasswordManager;
-
-public class PasswordDatabase {
-    private Connection connection;
-
-    public PWDatabase(String dbPath) throws SQLException {
-        connect(dbPath);
-    }
-
-    private void connect(String dbPath) throws SQLException {
-        connection = DriverManager.getConnection("jdbc:sqlite:" + dbPath);
-    }
-
-    //Method to add a password entry to the database
-    public void addPasswordEntry(PasswordEntry entry) {
-        // Implement database logic here
-    }
-
-    // Method to retrieve a password entry from the database
-    public PasswordEntry getPasswordEntry(String url) {
-        // Implement database logic here
-        return null;
-    }
-
-    //Method to update a password entry in the database
-    public void updatePasswordEntry(PasswordEntry) {
-        // Implement database logic here
-    }
-
-    public void close() throws SQLException {
-        if (connection != null && !connection.isClosed()) {
-            connection.close();
-        }
-    }
-
-
-    
-}
